Remove ignoring missing fields as a deserializer option. This is a Bad Idea(tm):
We cannot really ignore missing fields when loading, because this means that should we load template AND item(s) of that template in the same sync, we could load assed up items if they load before their template
(e.g. we'd lose field values until the next sync due to them not yet existing in SC)

Working on a fix that solves both problems.