Convert NameHint comment to a map, and parse the value
From a purely tactical standpoint this enables us to save the namehints when using U3.1's partial save values feature (since that only changes what the ItemChanges says was changed and leaves the other values alone).
From a longer term standpoint it enables gathering additional metadata about fields programmatically that you previously could visually.

For the record, DO NOT DEPEND ON THIS **HINT** VALUE BEING A CORRECT NAME FOR THE FIELD. IT IS **NOT** KEPT UPDATED IF YOU RENAME A FIELD. THIS IS FOR YOUR REFERENCE ONLY, DO NOT PROGRAM TO IT.