Proper Unversioned field support
Previously some unversioned support existed - by accident - and it treated them as versioned fields. This worked e.g. for display names.

This commit (along with its counterpart on Unicorn) brings first class support for Unversioned fields. This IS a breaking change for the YAML format as it nestles unversioned fields under the language nodes. It should be backwards compatible, at least mostly, but older versions of Rainbow will be unable to read the new unversioned fields and will throw an error.