Remove incorrect assumption about resetting standard values
Previous code could ignore setting a standard value just because the value in Sitecore was blank without a standard value. Which could mean that a standard value exists but the field value is explicitly set to blank. If the field is not present in the serialized item that does not mean the field should be blank, but that it should be the standard value.